+++ config/api/item-transfer/validation.php
@@ -5,7 +5,7 @@ declare(strict_types=1);
 return [
     'POST' => [
         'messages' => [
-            'resource_id.exists' => 'item-move/validation.resource_id-exists'
+            'resource_id.exists' => 'item-transfer/validation.resource_id-exists'
         ]
     ]
 ];