+++ config/api/request-error-log/validation.php
@@ -9,6 +9,7 @@ return [
             'expected_status_code' => 'required|integer|between:100,530',
             'returned_status_code' => 'required|integer|between:100,530',
             'request_uri' => 'required|string',
+            'source' => 'required|string|in:website,api,legacy,postman'
         ],
         'messages' => []
     ]