+++ resources/lang/en/route-descriptions.php
@@ -50,13 +50,13 @@ return [
     'item_sub_category_PATCH' => 'Update the subcategory assigned to the selected item',
     'item_sub_category_DELETE' => 'Delete the subcategory assigned to the selected item',
 
-    'item_move_POST' => 'Move an item to another resource',
+    'item_transfer_POST' => 'Transfer an item to another resource',
 
     'request_GET_access-log' => 'Return the access log, read requests',
     'request_GET_error_log' => 'Return the error log',
     'request_POST' => 'Create an error log report',
 
-    'summary_GET_request_access-log_monthly' => 'Return a summary of the access log, read requests, grouped by month',
+    'summary_GET_request_access-log' => 'Return a summary of the access log, all read requests, grouped by year and month',
 
     'summary-resource-type-GET-index' => 'Return a summary of the resource types',
     'summary-resource-GET-index' => 'Return a summary of the resources',